==> temp/13.training지역주민이찾는.py <==
from gensim.models import Word2Vec

# 기존 모델 불러오기
model_path = '../app/assets/modelfromreviews2.model'
model = Word2Vec.load(model_path)

# 새로운 문장 또는 문장 리스트
new_sentences = ["바닷가의 특성으로 회를 먹어야한다는 진부한 생각을 깨고자!! 오늘은 지역 주민이 더 많이 찾는다는 식당을 소개합니다.",
                 "이번 리플릿을 통해 강진 지역주민이 즐겨 찾는 맛집도 방문해 보시며 강진의 다양한 매력을 만나보세요~",
                 "지역 주민이 먹어야 찐 맛집?",
                 "이런 말이 있죠. 그 지역 주민들이 찾는 곳이 정말 인생 맛집이라는 얘기.",
                 "겉으로 보이는 것보다 정말 지역 주민들의 인정을 받은 곳을 좋아하시는 분들께는 좋은 소식입니다.",
                 "내 고향 손두부는 타지역 손님들보다 지역 사람들이 더 많이 찾는 맛집이에요~ ",
                 "이번 포스팅에서는 동두천지역 주민들이 자주 찾는 인기 맛집 TOP10을 소개하려고 합니다. 1",
                 "지역주민이 찾는 찐 수제우동집·코스모스 약국(마트)털이·소프트콘·타코야끼",
                 "덕계 지역주민만 찾는 맛집 숨은 맛집이네요 .",
                 "시원한 막걸리 잔술도 마시고 기부도 같이 할 수 있는 영통구청맛집 팔청식당 본점 역시 주변 주민들이 찾는 맛집의 다르네요 이 글은 리뷰노트를 통하여 본업체에서 서비스를 제공",
                 "지역 주민들이 찾는 진정한 숨은 맛집 센토칸 센즈역에서 가깝고 다양한 메뉴와 가정식다운 맛을 느낄 수 있답니다 센토칸(先頭館) 주소",
                 "지역주민들이 즐겨찾는 오래된 칼국수집 초읍 #산청칼국수",
                 "작고 아담한 매장이지만 늘 많은 손님들로 붐비는 지역 주민들이 많이 찾는 유명 맛집으로 통하는 곳이죠. 봄, 여름, 가을 날에는 야외 테이블에서 족발 & 소주를 즐길 수 있어 더욱 ",
                 "벚꽃 가로수가 꽤 길어서 벚꽃시즌에 왔으면 너무 멋있었을 것 같더라고요~ 내년에 이 길로 드라이브하기로~~ ㅎㅎ 순천 지역 주민이 찾는 향미정은 꼬막정식과 짱뚱어탕이 유명하다고 해요.",
                 "밑에 사진에 메뉴있으니 걱정마시구~ 지역주민이 찾는 무창포중국집 국빈원에 대해서 이야기 해볼께요~!! 국빈원가게 안에는 여자 남자화장실이 따로 되어있어요~ 다른 식당가면 화장",
                 "몇년전까지만 해도 지역주민만 알고 찾는 곳이었는데 이젠 멀리서도 오는듯 해서 아쉽더라구요. ㅎㅎㅎ 메뉴는 바지락 칼국수에요. 원래는 평택",
                 "조그마한 표선리 마을 사무소 근처에 지역주민이 많이 찾는 특색 있고 소박한 밥집이 많이 있었어요. 가시식당은 두루치기와 몸국 전문점",
                 "지역주민 찾는 제주 두루치기 맛집 가미돈 제주도 여행을 가면 고기국수나 흑돼지만 먹는줄 아는 분들 많으시던데 아닙니다.",
                 "오늘 소개하고 싶은 식당은 표선 지역주민들이 많이 찾는 임숙현 소머리 국밥 식당을 소개해볼까 합니다. 지역에 있는 주민들이 많이 찾는다는 것은 맛이",
                 "이름은 '창선반점'입니다 지역주민이 많이 찾는 남해맛집 쟁반짜장이 맛있는 남해 창선반점 실내가 깔끔하더군요 테이블은 10테이블 정도? 있네요 점심시간에 갔",
                 "그만큼 지역주민들이 찾는 진짜 #로컬맛집 동이손만두는 #국립수목원 (구 #광릉수목원 )과 #고모리저수지 #고모리카페거리 사이에 위치해있다. ",
                 "산들바람 영월은 정선과 더불어 곤드레밥이 유명한 지역입니다 영월에는 읍내에 유명한 곤드레밥집이 있고 많은 분들이 찾으시는데 물론 그 곳도 맛이 있지만 영월 지역주민이 찾아가는 맛집은 따"]

# 특정 단어
specific_word = "지역주민이찾는"

# 모델 업데이트
model.build_vocab(new_sentences, update=True)
model.train(new_sentences, total_examples=model.corpus_count, epochs=model.epochs)

# 특정 단어 관련 추가 학습
model.train([specific_word], total_examples=model.corpus_count, epochs=model.epochs)

# 업데이트된 모델 저장
model.save('updated_model.model')
