==> temp/19.Eu.py <==
import numpy as np

def euclidean_distance(vector_a, vector_b):
    if len(vector_a) != len(vector_b):
        raise ValueError("Vector dimensions do not match.")
    return np.linalg.norm(np.array(vector_a) - np.array(vector_b))

# 주어진 벡터

vector_a = [0.013375403, -0.053540263, -0.04729247, 0.032315426, 0.025813121, 0.01681236, 0.03607248, 0.010879169, -0.03589855, -0.006970225, 0.039399624, -0.058195062, -0.032164782, 0.05763016, 0.008390764, -0.021619251, -0.00093401666, -0.0077047227, -0.019975562, 0.016558537, -0.015192771, 0.017874677, 0.026840594, 0.013084967, 0.007649424, 0.19287968, -0.048521154, 0.054299694, -0.0064033708, 0.026451424, -0.048960518, -0.003043714, -0.015942276, -0.007414584, 0.015161903, 0.019654008, -0.014089825, -0.027781729, -0.0027100714, 0.027083404, 0.032554578, 0.039597277, 0.017072951, -0.015780628, -0.016479151, 0.07245805, -0.004294914, 0.035878256, -0.007694882, -0.05538574, 0.018851982, -0.03175608, -0.013128501, 0.03784488, 0.0096840635, 0.0043130144, -0.0075405645, 0.03504495, -0.010320883, -0.015212249, -0.0051684678, 0.062711984, -0.04841317, -0.024713388, 0.015143486, 0.008953076, -0.00548104, 0.03736658, 0.016126573, -0.04074153, -0.026385807, -0.005916508, -0.0531491, -0.03804127, 0.040613007, 0.023651373, 0.016375363, 0.037232406, -0.072308116, 0.03966164, 0.040133916, 0.019216724, -0.016036596, 0.005175424, -0.024299148, -0.0063452004, 0.027909627, 0.016436152, -0.023111029, -0.015954915, -0.018997869, -0.08423716, -0.011537101, 0.073278226, -0.026807336, -0.048625086, 0.026518013, 0.033167306, 0.023625657, -0.030773496, 0.018754505, 0.0024273756, 0.021314407, -0.03283113, 0.04768637, -0.08768238, -0.013493185, 0.011539216, -0.025962729, -0.030249659, -0.006339024, -0.031030333, -0.0037667495, 0.019998265, -0.03925155, -0.017477408, -0.017801339, -0.005085094, 0.0601044, 0.011468636, 0.04713033, -0.031462774, 0.010256181, 0.016001943, -0.026811535, 0.0021047313, -0.020121105, -0.022177504, 0.0047136126, -0.0095239375, 0.035260584, -0.0076265624, 0.029455377, 0.06572602, -0.0038693137, -0.023952955, 0.049861867, 0.06274906, 0.044251442, -0.022598172, 0.012581822, 0.012287395, -0.019099243, -0.042429473, 0.017240018, 0.04378037, 0.02852702, -0.06543241, -0.021296795, 0.04892192, 0.00094121695, 0.008397446, -0.02843461, -0.03284609, 0.001488436, -0.009854649, -0.02867983, 0.026858307, -0.034963757, 0.004924017, -0.040110663, 0.031693995, 0.0020189108, -0.02728196, 0.047044735, 0.041262425, 0.0454847, 0.005009885, -0.0076536597, 0.027692039, -0.03456588, -0.021756245, 0.013642835, 0.06362675, -0.03730592, 0.026438043, -0.025929917, 0.047354434, 0.019875063, -0.00451606, 0.03022144, 0.08319813, 0.00857477, 0.040689703, 0.037466254, 0.0018539075, -0.015246039, -0.02210506, 0.028396288, -0.037246536, 0.03725586, -0.013277752, -0.021513496, 0.00036065467, 0.012709285, 0.019780962, 0.053678256, 0.037907045, -0.026568843, 0.028699825, -0.0026882496, 0.026365558, 0.024272282, -0.039202493, -0.016594468, 0.0026466227, 0.041285332, 0.016794004, 0.027586622, -0.012083061, 0.014552617, 0.066721335, 0.020826317, 0.035035383, 0.004292568, 0.012410146, 0.015028632, -0.07979176, -0.011078206, 0.043422654, 0.012867599, 0.04189946, -0.027987063, 0.055073477, 0.043948628, 0.016250525, -0.002960261, 0.016644223, -0.04201327, 0.022979856, 0.008023962, -0.00021747313, 0.015195638, 0.021190997, -0.05138257, 0.0039011945, -0.02099637, 0.042670302, 0.0051646903, -0.017388446, 0.018137356, -0.00971882, 0.03922602, -0.055810846, 0.038340792, -0.016098581, 0.006883067, 0.059779495, 0.037133627, -0.034415178, -0.0002454843, 0.03143528, -0.033075556, 0.028314961, -0.008508221, -0.0008082781, -0.058934167, 0.02945596, 0.013638021, 0.019614145, -0.033822447, -0.0021414384, 0.00075605046, -0.09410386, -0.026603587, -0.04923242, 0.021407971, 0.008469304, -0.039588977, -0.05964351, 0.032696985, 0.023752932, -0.0127318185, -0.0015714439, -0.048626594, -0.055028338, -0.0637046, -0.024581704, -0.059466004, -0.026338425, 0.037483, 0.033196077, -0.079174206, 0.0027547712, 0.022337308, 0.0513945, -0.016747776, 0.06635308, 0.009853118, 0.012302158, -0.007194246, 0.025338106, -0.0029402953, 0.034980968, -0.030036565, 0.0042931996, -0.049920693, 0.0015296033, 0.0005254338, -0.016715612]
vector_b = [0.0043744706, -0.022709334, -0.0790146, 0.04190074, 0.05522604, 0.055816934, 0.03142039, -0.029665746, -0.07062997, -0.010239208, -0.01435435, -0.050120287, -0.077042684, 0.047214516, 0.051693257, 0.0027812794, -0.0043609217, 0.0064514955, -0.03178085, 0.032012194, -0.023763902, 0.06968558, -0.0006823316, 0.07265413, -0.016853405, 0.21977775, 0.028663827, 0.10727225, -5.6451914e-05, 0.017885748, -0.07206319, -0.060225546, -0.005303004, -0.0019226959, 0.003716324, 0.05586318, 0.011570986, -0.026275512, -0.062181156, 0.024955569, 0.0052939197, 0.026397191, -0.036591798, -0.07706517, -0.030646404, 0.12795907, 0.02763826, 0.123421565, -0.0038995906, -0.00075632334, -0.03867995, -0.048487693, -0.010977928, 0.004331777, -0.013418837, 0.059708133, 0.032585703, 0.087868795, -0.068066455, -0.039723784, -0.059513707, 0.0054354686, 0.030950006, 0.013197718, 0.034510724, 0.019034836, 0.0163195, 0.058407884, -0.06000546, 0.013903093, -0.0058460487, 0.00048468634, -0.110820204, 0.0058846436, 0.029045168, 0.10626153, -0.01114648, 0.076699235, -0.009082612, -0.0108070485, 0.0032380875, 0.0010621483, -0.056020595, -1.3029203e-06, -0.021370659, 0.061227925, -0.021248277, -0.02394618, 0.091276854, 0.021579023, -0.081578165, -0.121247426, 0.006129953, 0.10317005, 0.018897701, -0.016575877, 0.0005998665, 0.027622567, 0.02059944, -0.061184518, -0.0016120728, 0.0072267633, 0.012608815, -0.02108179, 0.044315666, -0.12849845, 0.04025945, 0.01008702, 0.027916864, -0.03206438, -0.04624718, -0.029037466, -0.011640683, 0.06857343, -0.0030196942, 0.017596986, -0.038828515, 0.030221725, 0.11004001, 0.013795551, 0.09681847, -0.048995394, -0.009639224, 0.068681434, -0.041463107, 0.02165909, -0.04517675, 0.023677731, 0.035489626, 0.014041927, -0.009770561, -0.01562264, 0.08421397, 0.06450428, -0.012449965, -0.05785697, 0.04034129, 0.077225536, 0.04017152, -0.018469531, 0.016046692, 0.008384958, -0.033432312, 0.009838071, 0.025043119, 0.055242106, 0.046281867, -0.0648017, -0.006499788, 0.062399425, 0.029388783, 0.013036994, 0.039530884, 0.017844697, 0.04859721, -0.019703876, -0.008468243, 0.0027723191, -0.035008177, -0.07608365, -0.07560471, -0.011024781, 0.015810689, -0.06488965, 0.04845745, 0.045040216, 0.0042098872, -0.031670604, -0.00023420947, 0.061465964, 0.008237125, -0.012581555, -0.013093073, 0.04561005, 0.01661971, 0.026777372, -0.021479687, 0.06275466, -0.009951698, -0.024411008, -0.0028026507, 0.10799897, -0.0044593057, 0.044403426, 0.017349178, -0.0019673528, -0.038674563, 0.010277728, 0.075296424, -0.02213642, 0.07936832, 0.054741535, -0.008119823, -0.09234103, -0.0073899827, 0.004720874, 0.06662337, 0.03925664, -0.04061038, 0.07512428, -0.031773206, -0.0014392193, 0.047881685, 0.04238979, -0.023339381, 0.03325495, 0.05941233, 0.038778372, 0.009272205, 0.006536734, -0.0137573, 0.08353386, 0.04880435, 0.021219049, -0.02745232, 0.018941425, 0.016866349, -0.028465766, 0.015314566, 0.0109832985, 0.013985634, 0.042029273, 0.017165545, 0.019290665, 0.097660296, 0.020722162, -0.004146926, -0.021676626, -0.07907562, -0.011133657, -0.026758444, -0.015184414, 0.041788276, 0.007756554, -0.07537488, 0.03403005, -0.009388132, 0.07546561, -0.028781787, -0.013838394, 0.0064789653, 0.012692868, 0.0644827, -0.07062146, 0.06283146, -0.016256042, 0.0013697324, 0.053282924, -0.014235273, -0.065296575, 0.0047800853, 0.08352697, -0.052422676, 0.05066974, 0.0053655664, 0.06338835, -0.07940154, -0.0014085248, 0.06470333, -0.02270728, -0.044157863, -0.019466698, 0.014955586, -0.04958364, 0.00034642778, -0.069548234, 0.07887743, 0.013784829, -0.002858702, -0.094287276, 0.060369935, -0.017191473, -0.01600504, 0.016684392, -0.049063414, -0.06837252, -0.06320358, -0.08359746, -0.020848196, -0.094327964, 0.031941433, 0.03769061, -0.12442723, -0.036839932, 0.0051387083, 0.041419342, 0.05016425, 0.117333055, 0.018595891, 0.05739522, -0.06260048, 0.04878528, -0.0088952845, 0.032592088, 0.036758997, 0.051970556, -0.018087912, 0.002874299, -0.0655896, -0.039760467]

# 유클리드 거리 계산
distance = euclidean_distance(vector_a, vector_b)

print(f"유클리드 거리: {distance}")
